Extends LinkedHashMap to fix #21
This isn't a real solution that should be used long term, but a means
to solve an issue for 0.2.0. The extended class should be removed
when we create a class for the payload.
What we're doing, is extending the LinkedHashMap class to check if
the value in the kv is null, so we avoid adding it into the map.